# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Learnova LMS API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /files:
    get:
      operationId: getFiles
      tags: [files]
      summary: List files for a subject (teacher or student)
      parameters:
        - in: query
          name: subject
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TeacherFile"
    post:
      operationId: uploadFile
      tags: [files]
      summary: Upload a file (teacher only)
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/FileUploadInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TeacherFile"
        "400":
          description: Invalid file type
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /files/{id}:
    delete:
      operationId: deleteFile
      tags: [files]
      summary: Delete a file (teacher only)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ── Students ──────────────────────────────────────────────────────────────
...
  /ratings/subject/{subject}:
    get:
      operationId: getMyRatingBySubject
      tags: [ratings]
      summary: Get student's rating for a specific subject
      parameters:
        - in: path
          name: subject
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Rating"
        "404":
          description: Not rated yet
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /ratings/{studentId}:
    put:
      operationId: upsertRating
      tags: [ratings]
      summary: Create or update a student rating (teacher only, upsert by student+subject)
      parameters:
        - in: path
          name: studentId
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RatingInput"
      responses:
        "200":
          description: Upserted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Rating"

  # ── Notebooks ─────────────────────────────────────────────────────────────
  /notebooks:
    get:
      operationId: getNotebooks
      tags: [notebooks]
      summary: List student's notebooks for a subject
      parameters:
        - in: query
          name: subject
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Notebook"
    post:
      operationId: createNotebook
      tags: [notebooks]
      summary: Create a notebook
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NotebookInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Notebook"
  /notebooks/{id}:
    patch:
      operationId: updateNotebook
      tags: [notebooks]
      summary: Rename a notebook
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NotebookUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Notebook"
    delete:
      operationId: deleteNotebook
      tags: [notebooks]
      summary: Delete a notebook
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── Notes ─────────────────────────────────────────────────────────────────
  /notes:
    get:
      operationId: getNotes
      tags: [notes]
      summary: List notes for a notebook
      parameters:
        - in: query
          name: notebookId
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Note"
    post:
      operationId: createNote
      tags: [notes]
      summary: Create a note
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NoteInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Note"
  /notes/{id}:
    patch:
      operationId: updateNote
      tags: [notes]
      summary: Update note content
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NoteUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Note"
    delete:
      operationId: deleteNote
      tags: [notes]
      summary: Delete a note
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── Note Files ────────────────────────────────────────────────────────────
  /note-files:
    get:
      operationId: getNoteFiles
      tags: [note-files]
      summary: List uploaded files for a notebook
      parameters:
        - in: query
          name: notebookId
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/NoteFile"
    post:
      operationId: uploadNoteFile
      tags: [note-files]
      summary: Upload a file to a notebook
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/NoteFileUploadInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NoteFile"
  /note-files/{id}:
    delete:
      operationId: deleteNoteFile
      tags: [note-files]
      summary: Delete a note file
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── AI ────────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]

    # ── Auth schemas ──────────────────────────────────────────────────────
...
    TeacherFile:
      type: object
      required: [id, filename, filepath, subject, uploadedBy, createdAt]
      properties:
        id:
          type: integer
        filename:
          type: string
        filepath:
          type: string
        subject:
          type: string
        uploadedBy:
          type: string
        createdAt:
          type: string
          format: date-time
    FileUploadInput:
      type: object
      required: [subject]
      properties:
        subject:
          type: string

    # ── Student schemas ───────────────────────────────────────────────────
...
    Rating:
      type: object
      required: [id, studentId, subject, teacherName, score, label, createdAt]
      properties:
        id:
          type: integer
        studentId:
          type: integer
        subject:
          type: string
        teacherName:
          type: string
        score:
          type: integer
          minimum: 0
          maximum: 100
        label:
          type: string
          enum: [Excellent, Good, Average, "Needs Improvement"]
        comment:
          type: ["string", "null"]
        studentName:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    RatingInput:
      type: object
      required: [score, comment]
      properties:
        score:
          type: integer
          minimum: 0
          maximum: 100
        comment:
          type: string

    # ── Notebook schemas ──────────────────────────────────────────────────
    Notebook:
      type: object
      required: [id, studentId, subject, notebookName, createdAt]
      properties:
        id:
          type: integer
        studentId:
          type: integer
        subject:
          type: string
        notebookName:
          type: string
        createdAt:
          type: string
          format: date-time
    NotebookInput:
      type: object
      required: [subject, notebookName]
      properties:
        subject:
          type: string
        notebookName:
          type: string
    NotebookUpdate:
      type: object
      required: [notebookName]
      properties:
        notebookName:
          type: string

    # ── Note schemas ──────────────────────────────────────────────────────
    Note:
      type: object
      required: [id, notebookId, content, createdAt]
      properties:
        id:
          type: integer
        notebookId:
          type: integer
        content:
          type: string
        createdAt:
          type: string
          format: date-time
    NoteInput:
      type: object
      required: [notebookId, content]
      properties:
        notebookId:
          type: integer
        content:
          type: string
    NoteUpdate:
      type: object
      required: [content]
      properties:
        content:
          type: string

    # ── Note File schemas ─────────────────────────────────────────────────
    NoteFile:
      type: object
      required: [id, notebookId, filename, filepath, createdAt]
      properties:
        id:
          type: integer
        notebookId:
          type: integer
        filename:
          type: string
        filepath:
          type: string
        createdAt:
          type: string
          format: date-time
    NoteFileUploadInput:
      type: object
      required: [notebookId]
      properties:
        notebookId:
          type: integer

    # ── AI schemas ────────────────────────────────────────────────────────
...