# lib/api-zod/src/generated/types/notebookInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Learnova LMS API
 * OpenAPI spec version: 0.1.0
 */

export interface NotebookInput {
  subject: string;
  notebookName: string;
}