# lib/api-zod/src/generated/types/ratingInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Learnova LMS API
 * OpenAPI spec version: 0.1.0
 */

export interface RatingInput {
  /**
   * @minimum 0
   * @maximum 100
   */
  score: number;
  comment: string;
}